requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.13.1
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
drf-yasg ==1.20.0
Pillow
